#16 #17 #18 処理分割
役割ごとに処理分割
